// frontend/src/api/generated/zod/budgets/budgets.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import { z as zod } from "zod";

export const getBudgetByIdParams = zod.object({
  id: zod.string().uuid(),
});

export const updateBudgetParams = zod.object({
  id: zod.string().uuid(),
});

export const updateBudgetBodyBudgetMonthRegExp = new RegExp(
  "^\\d{4}-(0[1-9]|1[0-2])$",
);
export const updateBudgetBodyBudgetAmountMin = 0.01;

export const updateBudgetBody = zod.object({
  categoryId: zod.string().uuid(),
  budgetMonth: zod.string().regex(updateBudgetBodyBudgetMonthRegExp),
  budgetAmount: zod.number().min(updateBudgetBodyBudgetAmountMin),
  autoRenew: zod.boolean(),
});

export const deleteBudgetParams = zod.object({
  id: zod.string().uuid(),
});

export const getBudgetsQueryParams = zod.object({
  year: zod.number().optional(),
  month: zod.number().optional(),
});

export const createBudgetBodyBudgetMonthRegExp = new RegExp(
  "^\\d{4}-(0[1-9]|1[0-2])$",
);
export const createBudgetBodyBudgetAmountMin = 0.01;

export const createBudgetBody = zod.object({
  categoryId: zod.string().uuid(),
  budgetMonth: zod.string().regex(createBudgetBodyBudgetMonthRegExp),
  budgetAmount: zod.number().min(createBudgetBodyBudgetAmountMin),
  autoRenew: zod.boolean(),
});